// src/services/store/action-type.ts
export enum DishAction {
	DISH = 'DISH',
	RESET_DISH = 'RESET_DISH',
	SELECT_NAME = 'SELECT_NAME',
	SELECT_TIME = 'SELECT_TIME',
	SELECT_DISH = 'SELECT_DISH',
	SELECT_DIAMETER = 'SELECT_DIAMETER',
	SELECT_PIZZA_SLICES = 'SELECT_PIZZA_SLICES',
	SELECT_SANDWITCH_SLICES = 'SELECT_SANDWITCH_SLICES',
	SELECT_SPICINESS_SCALE = 'SPICINESS_SCALE'
}